Improve comments for comments
Either --- or *** form line of block comments
if you can --- followed by spaces then a ( is is a block comment
and it will end at the matching closing parentheses ). In block comments
you need to have matching amounts of parentheses like
--- (
	example () () (())
)
this is a valid comment and
---(Comment) red 1 + 1 .
The last part of the code will be executed so it should have color as
well.